Revert "feat(sells): veículo na venda direta de oficina — seletor + cadastro rápido + placa na consulta (ADR 0251) (#2276)"
This reverts commit 01b7fa803b91aaab942dfd114119f250d06caa4d.
