+++ tests/Feature/Sells/SellsCreatePageTest.php
@@ -421,46 +421,3 @@ it('ProductSearchAutocomplete tipa campos backend explícitos (anti AP-12 F3)', 
     expect($componentSource)->toContain("type?: 'variable' | 'single'");
     expect($componentSource)->toContain('variation_group_price?: number');
 });
-
-// ─── ADR 0251 — Veículo na venda direta de oficina ──────────────────────────
-
-it('Page tem seletor de veículo gated por OficinaAuto (ADR 0251)', function () {
-    $source = readPage();
-    // Gate per-business — vestuário/ROTA LIVRE (hasOficinaAuto=false) não vê.
-    expect($source)->toContain('hasOficinaAuto');
-    // Plaquinha Mercosul reusada do OficinaAuto via shared.
-    expect($source)->toContain("from '@/Components/shared/MercosulPlate'");
-    expect($source)->toContain('<MercosulPlate');
-    // Consome o catálogo de veículos do cliente + liga em vehicle_id.
-    expect($source)->toContain('customerVehicles');
-    expect($source)->toContain('vehicle_id');
-    // Cadastro rápido sem perder a venda.
-    expect($source)->toContain('Cadastrar veículo');
-    expect($source)->toContain('QuickAddVehicleSheet');
-});
-
-it('QuickAddVehicleSheet faz fetch direto (não Inertia) pra não perder a venda (ADR 0251)', function () {
-    $src = file_get_contents(
-        base_path('resources/js/Pages/Sells/_components/QuickAddVehicleSheet.tsx'),
-    );
-    // POST no endpoint canônico de veículo do OficinaAuto.
-    expect($src)->toContain("fetch('/oficina-auto/veiculos'");
-    // CSRF manual (sem Inertia router) — preserva o draft da venda.
-    expect($src)->toContain('X-CSRF-TOKEN');
-    expect($src)->toContain('csrf-token');
-    // Placa + tipo obrigatórios (espelha StoreVehicleRequest).
-    expect($src)->toContain('vehicle_type');
-    expect($src)->toContain('contact_id');
-});
-
-it('Show e Index exibem a placa do veículo na consulta (ADR 0251)', function () {
-    $show = file_get_contents(base_path('resources/js/Pages/Sells/Show.tsx'));
-    expect($show)->toContain("from '@/Components/shared/MercosulPlate'");
-    expect($show)->toContain('headline.vehicle');
-
-    $tabela = file_get_contents(
-        base_path('resources/js/Pages/Sells/_components/SellsTabelaUnificada.tsx'),
-    );
-    expect($tabela)->toContain('vehicle_plate');
-    expect($tabela)->toContain('<MercosulPlate');
-});
